refactor: Rename `swapChannels` -> `editChannelPositions`.
